Move the WindowTitle component to /app
