+++ src/test/components/WindowTitle.test.js
@@ -8,7 +8,7 @@ import React from 'react';
 import { Provider } from 'react-redux';
 import { render } from '@testing-library/react';
 
-import { WindowTitle } from '../../components/shared/WindowTitle';
+import { WindowTitle } from 'firefox-profiler/components/app/WindowTitle';
 import {
   getEmptyProfile,
   getEmptyThread,
